linux revision

+++ src/geometry/bezier.cpp
@@ -8,16 +8,16 @@ BezierCurve::BezierCurve(int bezier_segments) : bezier_segments(bezier_segments)
 }
 
 void BezierCurve::DrawControlPoints(Color color_points, Color color_lines) const {
-    for (int i = 1; i < control_points.size(); ++i) {
+    for (int i = 1; i < static_cast<int>(control_points.size()); ++i) {
         DrawLineDotted(control_points[i - 1], control_points[i], 20, 3, color_lines);
     }
-    for (int i = 0; i < control_points.size(); ++i) {
+    for (int i = 0; i < static_cast<int>(control_points.size()); ++i) {
         DrawCircleV(control_points[i], 7, color_points);
     }
 }
 
 void BezierCurve::DrawCurve(Color color) const {
-    for (int i = 1; i < curve_points.size(); ++i) {
+    for (int i = 1; i < static_cast<int>(curve_points.size()); ++i) {
         DrawLineV(curve_points[i - 1], curve_points[i], color);
     }
 }
